# docs/swagger.yaml
basePath: /
definitions:
  handlers.OTPRequest:
    properties:
      phone:
        type: string
    type: object
  handlers.OTPVerifyRequest:
    properties:
      otp:
        type: string
      phone:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for sending and verifying OTP codes
  title: OTP Verification API
  version: "1.0"
paths:
  /otp/request:
    post:
      consumes:
      - application/json
      description: Send an OTP to the user's phone
      parameters:
      - description: Phone number
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.OTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request an OTP
      tags:
      - OTP
  /otp/verify:
    post:
      consumes:
      - application/json
      description: Verify the OTP sent to the user's phone and return a JWT token
        if valid
      parameters:
      - description: Phone number and OTP code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.OTPVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid or expired OTP
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server or Redis/DB error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify OTP
      tags:
      - OTP
schemes:
- http
swagger: "2.0"